backend/requirements.txt
Django==4.0.6
django-object-actions==3.1.0
djangorestframework==3.13.1
psycopg2>=2.8